feat(brochure): check isAdmin

+++ app/controllers/user/brochure.controller.js
@@ -1,6 +1,7 @@
 import { prisma } from "../../lib/dbConnect.js";
 import { verifyToken } from "../../lib/tokenHandler.js";
 import { removeFile } from "../../lib/multerStorage.js";
+import { getUserById } from "../../lib/userHandler.js";
 
 export const createBrochure = async (req, res, next) => {
   try {
@@ -78,35 +79,44 @@ export const updateBrochure = async (req, res, next) => {
     const data = verifyToken(req.headers.access_token);
     if (data?.status) return res.status(data.status).json(data);
 
+    const user = await getUserById(data.id);
+
     const brochureId = parseInt(req.params.id);
-    const brochure = await prisma.brochures.findUnique({
-      where: { id: brochureId, user_id: data.id },
-    });
-
-    if (!brochure) {
-      return res.status(404).json({
-        status: 404,
-        message: "Brochure not found.",
-      });
-    }
-
-    const { title } = req.body;
-    const images = req.files ? req.files.map((file) => file.path) : [];
-
-    const updatedBrochure = await prisma.brochures.update({
+    const brochure = await prisma.brochures.findFirst({
       where: { id: brochureId },
-      data: {
-        title,
-        images: {
-          create: images.map((imagePath) => ({ imagePath })),
-        },
-      },
-      include: {
-        images: true,
-      },
     });
 
-    res.json({ success: true, data: updatedBrochure });
+    if (!user.isAdmin) {
+      return res.status(401).json({
+        status: 401,
+        message: "Unauthorized: Access is denied due to invalid credentials.",
+      });
+    } {
+      if (!brochure) {
+        return res.status(404).json({
+          status: 404,
+          message: "Brochure not found.",
+        });
+      }
+
+      const { title } = req.body;
+      const images = req.files ? req.files.map((file) => file.path) : [];
+
+      const updatedBrochure = await prisma.brochures.update({
+        where: { id: brochureId },
+        data: {
+          title,
+          images: {
+            create: images.map((imagePath) => ({ imagePath })),
+          },
+        },
+        include: {
+          images: true,
+        },
+      });
+
+      res.json({ success: true, data: updatedBrochure });
+    }
   } catch (error) {
     next(error);
   }
@@ -117,30 +127,41 @@ export const deleteBrochure = async (req, res, next) => {
     const data = verifyToken(req.headers.access_token);
     if (data?.status) return res.status(data.status).json(data);
 
+    const user = await getUserById(data.id);
+
     const brochureId = parseInt(req.params.id);
     const brochure = await prisma.brochures.findUnique({
-      where: { id: brochureId, user_id: data.id },
-    });
-
-    if (!brochure) {
-      return res.status(404).json({
-        status: 404,
-        message: "Brochure not found.",
-      });
-    }
-
-    const deletedBrochure = await prisma.brochures.delete({
       where: { id: brochureId },
-      include: {
-        images: true,
-      },
     });
 
-    deletedBrochure.images.forEach((image) => {
-      removeFile(image.imagePath);
-    });
+    if (!user.isAdmin) {
+      return res.status(401).json({
+        status: 401,
+        message: "Unauthorized: Access is denied due to invalid credentials.",
+      });
+    } {
 
-    res.json({ success: true, data: deletedBrochure });
+
+      if (!brochure) {
+        return res.status(404).json({
+          status: 404,
+          message: "Brochure not found.",
+        });
+      }
+
+      const deletedBrochure = await prisma.brochures.delete({
+        where: { id: brochureId },
+        include: {
+          images: true,
+        },
+      });
+
+      deletedBrochure.images.forEach((image) => {
+        removeFile(image.imagePath);
+      });
+
+      res.json({ success: true, data: deletedBrochure });
+    }
   } catch (error) {
     next(error);
   }
